% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExplAnalysis_fn.R
\name{ExploratoryAnalysis}
\alias{ExploratoryAnalysis}
\title{Exploratory Analysis}
\usage{
ExploratoryAnalysis(data = NULL)
}
\arguments{
\item{\code{data}}{a processed data frame of ion's concentrations (columns) for each gene's knockout (rows) (wide format) e.g. \code{data.wide} from the \code{PreProcessing_fn}}
}
\value{
\code{plot.Pearson_correlation} plot of Pearson's correlations across ions

\code{plot.PCA_Individual} plot of ion's profiles in a lower-dimensional space from PCA

\code{plot.heatmap} plot of ion's concentrations for each gene's knockout with added clustering dendogram

\code{plot.pairwise_correlation_map} plot of pairwise correlation coefficients across ions

\code{plot.regularized_partial_correlation_network} plot of partial correlation coefficients for pairs of ions after conditioning for the remaining ions. Red for negative, green for positive partial correlations. No connections drawn when partial correlations equals 0

\code{data.PCA_loadings} a data frame of weights of each of the original gene knockouts also known as loading vectors associated to each PC
}
\description{
Exploratory tools for Ionomics data
}
\examples{
\code{
library(IonFlow)

### Run PreProcessing function
pre_proc <- PreProcessing(data=IonData,stdev=pre_defined_sd)

### Run ExploratoryAnalysis function
exp_anal <- ExploratoryAnalysis(data=pre_proc$data.wide)

# plots
exp_anal$plot.Pearson_correlation
exp_anal$plot.PCA_Individual
exp_anal$plot.heatmap
exp_anal$plot.pairwise_correlation_map
exp_anal$plot.regularized_partial_correlation_network

# data
head(exp_anal$data.PCA_loadings)
}
}
