% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClustering_fn.R
\name{GeneClustering}
\alias{GeneClustering}
\title{Clustering}
\usage{
GeneClustering(data = NULL, data_Symb = NULL)
}
\arguments{
\item{\code{data}}{a processed data frame of ion's concentrations (columns) for each gene's knockout (rows) (wide format) e.g. \code{data.wide} from the \code{PreProcessing_fn}}

\item{\code{data_Symb}}{a processed data frame of symbolised (values -1,0, or 1) ion's concentration profiles (columns) for each gene's knockout (rows) (wide format) e.g. \code{data.wide_Symb} from the \code{PreProcessing_fn}}
}
\value{
\code{stats.clusters statistics} of the gene's count by cluster

\code{stats.Kegg_Goslim_annotation} statistics from the GO Slim annotation

\code{stats.Goterms_enrichment} statistics from the GO terms enrichment

\code{plot.profiles} plot of the gene's profiles for each cluster
}
\description{
Clustering and enrichment tools for Ionomics data
}
\examples{
\code{
library(IonFlow)

### Run PreProcessing function
pre_proc <- PreProcessing(data=IonData,stdev=pre_defined_sd)

### Run GeneClustering function
gene_clust <- GeneClustering(data=pre_proc$data.wide, data_Symb=pre_proc$data.wide_Symb)

# stats
gene_clust$stats.clusters
gene_clust$stats.Kegg_Goslim_annotation
gene_clust$stats.Goterms_enrichment

# plots
gene_clust$plot.profiles
}
}
