% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing_fn.R
\name{PreProcessing}
\alias{PreProcessing}
\title{Pre-processing}
\usage{
PreProcessing(data = NULL, stdev = NULL)
}
\arguments{
\item{\code{data}}{a data frame of ion's concentrations (columns) for various gene's knockout (rows)}

\item{\code{stdev}}{an optional pre-defined vector of standard deviations for each ion}
}
\value{
\code{stats.raw_data} statistics for the raw data ion concentration values

\code{stats.outliers} statistics (raw count and percentage of outliers) from the raw data for each ion

\code{stats.median_batch_corrected_data} statistics for the median batch corrected data

\code{stats.standardised_data} statistics for standardised data

\code{plot.logConcentration_z_scores} plot of the symbolised ion concentration profiles of the knockouts

\code{plot.logConcentration_by_batch} plot by batch of the log-ion concentration values after outlier removal and median batch correction

\code{dataR.long} a processed data frame of ion's concentrations with batch-replicated gene's knockout (long format)

\code{data.long} a processed data frame of ion's concentrations for each gene's knockout (long format)

\code{data.wide} a processed data frame of ion's concentrations (columns) for each gene's knockout (rows) (wide format)

\code{data.wide_Symb} a processed data frame of symbolised (values -1,0, or 1) ion's concentration profiles (columns) for each gene's knockout (rows) (wide format)
}
\description{
Pre-processing tools for Ionomics data
}
\examples{
\code{
library(IonFlow)

### Run PreProcessing function
pre_proc <- PreProcessing(data=IonData,stdev=pre_defined_sd)

# stats
pre_proc$stats.raw_data
pre_proc$stats.outliers
pre_proc$stats.median_batch_corrected_data
pre_proc$stats.standardised_data

# plots
pre_proc$plot.logConcentration_by_batch
pre_proc$plot.logConcentration_z_scores

# data
head(pre_proc$dataR.long)
head(pre_proc$data.long)
head(pre_proc$data.wide)
head(pre_proc$data.wide_Symb)
}
}
